전체적인 맵의 틀, 몬스터 스탯 데이터 제작 아이템 프리팹 생성 경험치, 데미지 로직 수정

+++ Assets/Eruption RPG/Scripts/DropTable.cs
@@ -15,7 +15,7 @@ public class DropItem
 public class DropTable : ScriptableObject
 {
     public int money;
-    public int exp;
+    public long exp;
     public int battlePoint;
     public DropItem[] dropItems;
